pyproject.toml
[tool.poetry]
name = "fakerhelp"
version = "0.1.0"
description = ""
authors = ["Karl Kroening <karlk@kralnet.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = {extras = ["colors"], version = "^5.12.0"}
mock = "^5.1.0"
mypy = "^1.4.1"
pylint = "^2.17.4"
pytest = "^7.4.0"
pytest-faker = "^2.0.0"
pytest-only = "^2.0.0"
pytest-randomly = "^3.13.0"
pytest-xdist = "^3.3.1"
types-mock = "^5.0.0.7"
types-setuptools = "^68.0.0.1"
ipython = "^8.14.0"

[tool.poetry.scripts]
fakerhelp = "fakerhelp:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '^/(fakerhelp|tests)/.*\.pyi?$'
skip-string-normalization = true
target_version = ['py311']

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = true
ignore_whitespace = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
no_sections = true
use_parentheses = true

[tool.mypy]
files = 'fakerhelp,tests'
mypy_path = 'stubs'
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
allow_redefinition = true
module = "tests.*"

[[tool.mypy.overrides]]
module = "fakerhelp.*"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
no_implicit_optional = true
no_implicit_reexport = true
strict_concatenate = true
strict_equality = true
warn_return_any = true
warn_unused_ignores = true
warn_unreachable = true

[tool.pytest.ini_options]
addopts = '''
    --pdbcls=IPython.terminal.debugger:Pdb
    -W "error::pytest.PytestUnhandledCoroutineWarning"
    -W "error::pytest.PytestUnraisableExceptionWarning"
    -W "error::RuntimeWarning"
'''
markers = []
testpaths = ["tests"]
